// src/app/(protected)/dashboard/meeting-and-track-targets/[meetingId]/status/started/form-modals/form-modal-types.ts
export enum TOPIC_TYPE {
  CHAIR_INTRODUCTION = 'chairs_introduction_and_opening',
  ATTENDEES = 'attendees_and_conflicts_of_interest',
  APPROVAL = 'approval_of_previous_minutes',
  ADDITIONAL_OPENING = 'additional_opening_item',
  RISKS = 'risks_for_review',
  HORIZON_PLANNING = 'horizon_planning',
  PROGRESS = 'progress_against_objectives',
  NEXT_MEETING = 'date_of_next_meeting',
  ADDITIONAL_AOB = 'additional_aob_item',
  ACTIONS = 'actions_from_meeting',
  REVIEW_APPROVAL = 'for_review_or_approval',
  RECORD_APPROVAL = 'record_for_review',
  REGISTER_APPROVAL = 'register_for_review',
  OBJECTIVE_APPROVAL = 'objective_for_review',
  PROGRESS_ACTIONS = 'progres_against_actions',
  FINANCE = 'finance_and_operations',
  ADDITIONAL_DISCUSSION = 'additional_discussion_item',
  RISK_REGISTER = 'risk_register',
  ADDITIONAL_INFORMATION = 'additional_information_item',
  COMPLAINT_INCIDENT = 'complaints_and_or_incidents',
  ADDITIONAL_TOPIC = 'additional_topic'
}

export enum CATEGORY_TYPE {
  DISCUSSION = 'for_discussion',
  AOB = 'any_other_business',
  ACTIONS_MEETING = 'actions_from_this_meeting',
  OPENING_MEETING = 'opening',
  INFORMATION = 'for_information',
  REVIEW_APPROVAL = 'for_review_or_approval',
  ACTION_TRACKER = 'action_tracker',
}

export enum MAIN_HEADING {
  DISCUSSION = 'For Discussion',
  AOB = 'Any other business (AOB)',
  ACTIONS_MEETING = '',
  OPENING_MEETING = 'Opening The Meeting',
  INFORMATION = 'For Information',
  REVIEW_APPROVAL = 'For Review / Approval',
  ACTION_TRACKER = 'Action Tracker',
  ACTIONS_FROM_PREVIOUS_MEETINGS = 'Actions from previous meeting(s)'
}

export enum SUB_HEADING {
  CHAIR_INTRODUCTION = 'Chair’s Introduction and Opening',
  ATTENDEES = 'Attendees & Conflicts of interest',
  APPROVAL = 'Approval of Previous Minutes',
  ADDITIONAL_OPENING = 'Additional Opening Item',
  RISKS = 'Risks For Review',
  HORIZON_PLANNING = 'Horizon Planning',
  PROGRESS = 'Progress Against Objectives',
  NEXT_MEETING = 'Next Meeting',
  ADDITIONAL_AOB = '',
  ACTIONS = '',
  REVIEW_APPROVAL = '',
  PROGRESS_ACTIONS = 'Progress against actions',
  ACTION_TRACKER = 'Action Tracker',
  FINANCE = 'Finance & Operations',
  ADDITIONAL_DISCUSSION = 'Additional Discussion Item',
  RISK_REGISTER = 'Risk Register',
  ADDITIONAL_INFORMATION = 'Additional Information Item',
  COMPLAINT_INCIDENT = 'Complaints and/or Incidents',
  ATTENDEES_MANAGEMENT = 'Attendees'
}